requirements.txt
accelerate==1.6.0
datasets==3.5.1
transformers==4.51.3
triton==3.2.0
trl==0.17.0
peft==0.15.2
deepspeed==0.16.7
hydra-core==1.3
lm_eval
lm_eval[japanese_leaderboard]